Add coccicheck
This is kind of a pain and I am not too pleased with it.

Essentially we want to run coccicheck and look for any warnings
that propogated from this pull request.

Annoyance #1: coccicheck takes a long time to run, we only care about
checking code from these new patches

Solution #1: Find out files touched by PR and only run coccicheck on
those. This means it should have Makefile in same dir... This also
means we should use a matrix to loop over all the areas that we touched

Annoyance #2: We only care about new issues from the pull request. This
means any issues that existed before this PR should be skipped.

Solution #2: Add another dimension to our matrix that generates a
coccicheck report before the PR, generate another report for output
with the PR applied.

Annoyance #3: We want our coccicheck matrix instances to run in
parallel. On overage, just 1 run of these over hfi1 takes around 5
minutes. But, we want to analyze the result of old vs new reports when
they finish

Solution #4: Add a new job that waits on our matrix job. This job is a
matrix of just our touched paths (from Solution #1). This job will
look at the results coccicheck

Annoyance #5: Transfering data between jobs is not easy with Github
Actions, especially if this data is from a dynamically sized matrix.
We need to give the coccicheck report to the next job for analysis.

Solution #5: Put the results of coccicheck in a cache. IOW make a cache
for every result. This is actually kind of nice because we can use the
sha of the commit to avoid rerunning coccicheck for commit's we already
generated.

Annoyance #6: Coccicheck has nonuniform output. I thought of just
running coccicheck for the HEAD commit and looking if any error
is from our PR just by parsing out file names and line numbers
from the error and using git blame to see if it is from our PR
commits but we cannot because the error line only sometimes
comes with a line number AND it only sometimes has a file!

Solution #6: We need output of before and after PR and
we need to be creative about diff'ing these reports.
Things like line numbers can change but that error could
have existed already.

Signed-off-by: Nick Child <nchild@cornelisnetworks.com>