Add smatch and build support
Since smatch relies on Modules.symvers, we need to build
the kernel now, while we are at it we can
upload rpm artifacts.
This makes the run take much longer (probably close to an hour
and a half).
To build we need a config file,  for now lets just assume
that the confif file exists in the workflow directory.

The build cache is a directory of all softlinks to the
actual source cache. This means there is a dependency from
build to source caches. Hopefully this makes our cache smaller.

There was an option to build the kernel when creating the original
source cache but I figured that would cause all other jobs to wait
until building was done to run. So I favored, parrallelism and getting
easy quick things done early.

Smatch jobs are very similar to coccinelle (go read that commit message
for an explanation on ugliness): We need to load kernel caches, we need
to run in a matrix for all touched areas, we run with and without the new
patchset, we save results to a new cache where another job can read and
parse results.

Signed-off-by: Nick Child <nchild@cornelisnetworks.com>